Moved some utilities in geo/utils.py
